Have added the database connector, which automatically fills all the tables
