--- postgresConnection/Readme.md
@@ -1,2 +1,0 @@
-Program to operate on postgres DB using a backend programming language. 
-Code for java and python are added
